+++ Isatronia/Exceptions/ImageException.h
@@ -1,3 +1,9 @@
+//--------------------------------------------------------------------------------------
+// File: ImageException.h
+//
+// Copyright (c) Ishgrina. All rights reserved.
+// Licensed under the MIT License.
+//--------------------------------------------------------------------------------------
 #pragma once
 #include <iostream>
 #include "./Exception.h"
