+++ Isatronia/Framework/Macro.h
@@ -1,3 +1,9 @@
+//--------------------------------------------------------------------------------------
+// File: Macro.h
+//
+// Copyright (c) Ishgrina. All rights reserved.
+// Licensed under the MIT License.
+//--------------------------------------------------------------------------------------
 
 //--------------------------------------------------------------
 // Macros about COM
