+++ Isatronia/Framework/Utils.cpp
@@ -1,0 +1,6 @@
+//--------------------------------------------------------------------------------------
+// File: Utils.cpp
+//
+// Copyright (c) Ishgrina. All rights reserved.
+// Licensed under the MIT License.
+//--------------------------------------------------------------------------------------
