+++ IsatroniaFrame/IsatroniaFrame.vcxproj.filters (2).bak
@@ -1,0 +1,3 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="Current" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+</Project>
